move DastReport struct
