+++ tests/rhwa/internal/rapidast/dastreport.go
@@ -1,0 +1,33 @@
+package rapidast
+
+type DASTReport struct {
+	ClusterName string
+	Resources   []struct {
+		Name      string
+		Namespace string
+		Results   []struct {
+			Target         string
+			Class          string
+			Type           string
+			MisconfSummary struct {
+				Success    int
+				Failures   int
+				Exceptions int
+			}
+			Misconfigurations []struct {
+				Type        string
+				ID          string
+				AVDID       string
+				Description string
+				Message     string
+				Namespace   string
+				Query       string
+				Resolution  string
+				Severity    string
+				PrimaryURL  string
+				References  []string
+				Status      string
+			}
+		}
+	}
+}
